delete vector_block_reader.hpp
