--- libs/iresearch/include/iresearch/formats/ivf/vector_block_reader.hpp
@@ -1,62 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-/// DISCLAIMER
-///
-/// Copyright 2026 SereneDB GmbH, Berlin, Germany
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     http://www.apache.org/licenses/LICENSE-2.0
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-/// Copyright holder is SereneDB GmbH, Berlin, Germany
-////////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#include <cstddef>
-#include <cstdint>
-#include <span>
-#include <utility>
-#include <vector>
-
-#include "basics/assert.h"
-#include "iresearch/store/data_input.hpp"
-#include "iresearch/types.hpp"
-
-namespace irs {
-
-class VectorBlockReader {
- public:
-  VectorBlockReader(IndexInput::ptr&& in, uint32_t record_size) noexcept
-    : _in{std::move(in)}, _record_size{record_size} {
-    SDB_ASSERT(_in);
-  }
-
-  void Reset(uint64_t base_offset) noexcept { _base = base_offset; }
-
-  std::span<const byte_type> Read(size_t index, size_t count) {
-    const uint64_t offset = _base + static_cast<uint64_t>(index) * _record_size;
-    const size_t bytes = count * size_t{_record_size};
-    if (const byte_type* p = _in->ReadVolatile(offset, bytes)) {
-      return {p, bytes};
-    }
-    _buf.resize(bytes);
-    _in->ReadData(offset, _buf.data(), bytes);
-    return _buf;
-  }
-
- private:
-  IndexInput::ptr _in;
-  std::vector<byte_type> _buf;
-  uint64_t _base = 0;
-  uint32_t _record_size;
-};
-
-}  // namespace irs
